+++ Changelog.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Bug fix in atomic/atom/startupinit at sqlite3 method. Completed on 2025-11-19
 - Bug fix in compmgr engine. Completed on 2025-11-25
+- Bug fix in deskelectronjs engine which is memory leakage cause by session frequence query and cause the entire application corrupted. Completed on 2025-11-25
 
 ### Security